Move navigation routes into feature layer
