+++ src/app/App.stories.tsx
@@ -8,7 +8,7 @@ import type { Meta, StoryObj } from "@storybook/react";
 import { expect } from "storybook/test";
 
 import { AppRoutes } from "@/app/routes";
-import { routes } from "@/shared/routes";
+import { routes } from "@/features/navigate";
 import { type PageStoryParameters, preparePageStory, withPageStory } from "@/storybook/PageDecorator";
 import { PAGE_STORY_CARD_ID, PAGE_STORY_DECK_ID, pageStoryState } from "@/storybook/pageFixture";
 import { STORYBOOK_DECK_IMPORT_URL } from "@/storybook/handlers";
