--- src/shared/routes/index.ts
@@ -1,2 +1,0 @@
-export { routes } from "./routes";
-export { useNavigation } from "./useNavigation";
